% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_events.R
\name{import_tei_events}
\alias{import_tei_events}
\alias{create_events}
\alias{generate_tei_events_payload}
\title{Import events in a DHIS2 tracker program}
\usage{
create_events(
  tei_events,
  baseurl = NULL,
  user = NULL,
  pass = NULL,
  ua = NULL,
  timeout = 60
)

generate_tei_events_payload(tei_events)
}
\arguments{
\item{tei_events}{A data.frame object, containing a list of events to import.}

\item{baseurl}{the baseurl.}

\item{user}{DHIS2 username.}

\item{pass}{DHIS2 password.}

\item{ua}{String, name that identifies the user agent.}

\item{timeout}{integer, specify the timeout in seconds.}
}
\description{
Imports single events into a tracker program through the \verb{DHIS2 events API}
\code{endpoint}.
}
\section{Functions}{
\itemize{
\item \code{create_events}: Create events in a DHIS2 program

\item \code{generate_tei_events_payload}: Generate events payload
}}

