% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tei_events.R
\docType{data}
\name{tei_events}
\alias{tei_events}
\title{Tracked entity instance events}
\format{
\describe{
\item{event}{event uids}
\item{status}{status of the event}
\item{program}{program id}
\item{programStage}{id of a program stage}
\item{enrollment}{enrollment id}
\item{orgUnit}{the orgUnit id}
\item{eventDate}{the event date}
\item{dueDate}{due date of the event}
\item{latitude}{coordianates, latitude}
\item{longitude}{coordianes, longitude}
\item{dataElement}{id of the datalement}
\item{value}{value of the data element}
\item{storedBy}{name of person / account uploading the events}
\item{providedElseWhere}{logical, is this dataElement provided elsewhere?}
\item{completedDate}{date of completion}
\item{completedBy}{person or account completing the event}
\item{geometry}{coordinates}
}
}
\source{
DHIS2 guide
}
\usage{
tei_events
}
\description{
A dataset containing tracked entity instance events for import.
}
\keyword{datasets}
